Allow disabling compatibility checks
Checks disabled by calling MastodonBuilder.disableCompatibilityChecks() include the initial check for the server software reporting as Mastodon while building the client, as well as subsequent checks for the exact API version in use. This allows the resulting MastodonClient to be used with other server software offering a Mastodon-compatible API, at the cost of having to check for invalid requests yourself. Endpoints and individual attributes not defined in the official Mastodon API are not supported.

Closes #577.