Further compatibility checks, including option to disable (#581)
* Allow disabling compatibility checks

Checks disabled by calling MastodonBuilder.disableCompatibilityChecks() include the initial check for the server software reporting as Mastodon while building the client, as well as subsequent checks for the exact API version in use. This allows the resulting MastodonClient to be used with other server software offering a Mastodon-compatible API, at the cost of having to check for invalid requests yourself. Endpoints and individual attributes not defined in the official Mastodon API are not supported.

Closes #577.

* Add version check while building MastodonClient

- introduce a minimumSupportedVersion, currently set to 4.2.0
- check against this minimum and throw UnsupportedServerException if connected server runs Mastodon at a lower version
- implement partial parsing of version strings according to semantic versioning

* Make SemanticVersion extend Comparable<SemanticVersion>

(also remove debugging code)

---------

Co-authored-by: Patrick Geselbracht <code@patrick-geselbracht.eu>